backend/requirements.txt
fastapi
uvicorn
gunicorn
uvloop
httptools
httpx
asyncpg
regex
pyyaml
python-dateutil
python-json-logger
pytz
pytest
pytest-cov
flake8
pytest-flake8
pytest-mypy
pytest-profiling
freezegun
async-timeout
yoyo-migrations
psycopg2
aiofiles
